+++ haystack/modeling/training/dpr.py
@@ -1,0 +1,1 @@
+# TODO create DPR_Trainer class here that can be called from retriever.dense.DPR.train()
