+++ haystack/modeling/training/question_answering.py
@@ -1,0 +1,1 @@
+# TODO make QA_Trainer class and use insider reader.train
