+++ haystack/modeling/visual/ascii/images.py
