+++ test/test_modeling_processor_saving_loading.py
@@ -1,9 +1,9 @@
 import logging
 from pathlib import Path
 
-from haystack.basics.data_handler.processor import SquadProcessor
-from haystack.basics.modeling.tokenization import Tokenizer
-from haystack.basics.utils import set_all_seeds
+from haystack.modeling.data_handler.processor import SquadProcessor
+from haystack.modeling.model.tokenization import Tokenizer
+from haystack.modeling.utils import set_all_seeds
 import torch
 
 
@@ -20,7 +20,7 @@ def test_processor_saving_loading(caplog):
 
     processor = SquadProcessor(
         tokenizer=tokenizer,
-        max_seq_len=128,
+        max_seq_len=256,
         label_list=["start_token", "end_token"],
         train_filename="train-sample.json",
         dev_filename="dev-sample.json",
@@ -29,14 +29,14 @@ def test_processor_saving_loading(caplog):
     )
 
     dicts = processor.file_to_dicts(file=Path("samples/qa/dev-sample.json"))
-    data, tensor_names, _ = processor.dataset_from_dicts(dicts)
+    data, tensor_names, _ = processor.dataset_from_dicts(dicts=dicts, indices=[1])
 
     save_dir = Path("testsave/processor")
     processor.save(save_dir)
 
     processor = processor.load_from_dir(save_dir)
     dicts = processor.file_to_dicts(file=Path("samples/qa/dev-sample.json"))
-    data_loaded, tensor_names_loaded, _ = processor.dataset_from_dicts(dicts)
+    data_loaded, tensor_names_loaded, _ = processor.dataset_from_dicts(dicts, indices=[1])
 
     assert tensor_names == tensor_names_loaded
     for i in range(len(data.tensors)):
